改变显示代码逻辑，页面显示逻辑优化，初步实现温控转速调整，不过OLED屏无法稳定显示，改善屏幕黑屏问题，但问题仍存在

+++ demo9-x 智能温控风扇系统/demo9-x1 串口发送温控信号+PWM占空比/Hardware/Motor.c
@@ -23,7 +23,7 @@ void Motor_Init(void)
 	
 }
 
-void Motor_SetSpeed(int8_t Speed)
+void Motor_SetSpeed(int16_t Speed)
 {
 	if(Speed >= 0)
 	{
@@ -47,9 +47,9 @@ void Motor_Turn(void)
 		OLED_ShowString(3,12,"   ");
 		OLED_ShowSignedNum(3,12,Speed,3);
 		if(GPIO_ReadOutputDataBit(GPIOA, GPIO_Pin_11) == 1)
-			{
-				Motor_SetSpeed(Speed);
-			}
+		{
+			Motor_SetSpeed(Speed);
+		}
 	}
 	else
 	{
